\encoding{latin1}
\name{LeLogicielR-package}
\alias{LeLogicielR-package}
\alias{LeLogicielR}
\docType{package}
\title{Package illustrant le livre: Le Logiciel R} 
\description{
Ce package permet d'utiliser certaines fonctions contruites à l'occasion
de l'écriture du livre :\emph{Le logiciel R, Maitriser le langage, Effectuer des analyses statistiques}.
Vous pouvez aussi y trouver les jeux de données utilisées dans ce livre. 
}
\details{
\tabular{ll}{
Package: \tab LeLogicielR\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2010-10-09\cr
License: \tab GPL(>=2.0)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Lafaye de Micheaux Pierre <lafaye@dms.umontreal.ca>, Remy Drouilhet <Remy.Drouilhet@upmf-grenoble.fr>, Liquet Benoit <benoit.liquet@isped.u-bordeaux2.fr>
}
\references{Livre: Le logiciel R, Maitriser le langage, Effectuer des analyses statistiques}
\keyword{package}


